Cell_Curvature_Measurements
Now measures the curvature for every Borrelia cell. This uses the medial axis and performs a bivariate low-degree polynomial fit to find the "trend" of the cells general shape. It then measures the residuals around this fit. The standard deviation of the residuals captures the variance of the residuals, or how "curvy" each cell is.